# server/models/Prmt_db/generated/ProjectsModelGenerated.js
/**
 * 
 * 
  _____                      _              _ _ _     _   _     _        __ _ _      
 |  __ \                    | |            | (_) |   | | | |   (_)      / _(_) |     
 | |  | | ___    _ __   ___ | |_    ___  __| |_| |_  | |_| |__  _ ___  | |_ _| | ___ 
 | |  | |/ _ \  | '_ \ / _ \| __|  / _ \/ _` | | __| | __| '_ \| / __| |  _| | |/ _ \
 | |__| | (_) | | | | | (_) | |_  |  __/ (_| | | |_  | |_| | | | \__ \ | | | | |  __/
 |_____/ \___/  |_| |_|\___/ \__|  \___|\__,_|_|\__|  \__|_| |_|_|___/ |_| |_|_|\___|
 
 * DO NOT EDIT THIS FILE!! 
 * 
 *  TO CUSTOMIZE projectsModelGenerated.js PLEASE EDIT ../projectsModel.js
 * 
 *  -- THIS FILE WILL BE OVERWRITTEN ON THE NEXT SKAFFOLDER'S CODE GENERATION --
 * 
 */
// Database
import Database from "../../../classes/Database_Prmt_db";
import mongoose, { Schema } from "mongoose";

// Logger
import Logger from "../../../classes/Logger";

const generatedModel = {
  /**
   * Init  schema
   */
  init() {
    const db = Database.getConnection();

    /**
      * projects
      */
    const projectsSchema = new mongoose.Schema({
      projApproval: {
        type: "Object"
      },
      projAreaAccess: {
        type: "String"
      },
      projCoordinator: {
        type: "String"
      },
      projCoordinatorEmail: {
        type: "String"
      },
      projCoordinatorTel: {
        type: "Number"
      },
      projDateSubmitted: {
        type: "Date"
      },
      projDesc: {
        type: "String"
      },
      projEndDate: {
        type: "Date"
      },
      projGeneralContractor: {
        type: "String"
      },
      projGeneralContractorSupervisor: {
        type: "String"
      },
      projGeneralContractorSupervisorEmail: {
        type: "String"
      },
      projGeneralContractorSupervisorTel: {
        type: "Number"
      },
      projManager: {
        type: "String"
      },
      projName: {
        type: "String"
      },
      projStartDate: {
        type: "Date"
      },
      projType: {
        type: "Object"
      },
      projectArea: {
        type: "String"
      },
      // RELATIONS
      
      
      // EXTERNAL RELATIONS
      /*
      */
    });

    generatedModel.setModel(db.connection.model("Projects", projectsSchema));

    return projectsSchema;
  },

  /**
   * Set Model
   */
  setModel: model => {
    generatedModel.model = model;
  },

  /**
   * Get model
   */
  getModel: () => {
    return generatedModel.model;
  },

  // Start queries
    

  // CRUD METHODS




};

export default generatedModel;